Rename edge file
